feat: implement navigation 2.0 in the whole app
