+++ lib/features/auth/presentation/pages/login_page.dart
@@ -25,7 +25,7 @@ class LoginPage extends StatelessWidget {
         listener: (context, state) {
           if (state is AuthenticatedWithLeetcodeAccount ||
               state is AuthenticatedWithLeetcodeUserName) {
-            context.router.replace(const DashboardRoute());
+            context.router.root.replace(const DashboardRoute());
           }
         },
         child: AppUpdater(
