+++ lib/features/auth/presentation/shells/auth_shell.dart
@@ -1,0 +1,12 @@
+import 'package:auto_route/auto_route.dart';
+import 'package:flutter/material.dart';
+
+@RoutePage()
+class AuthShell extends StatelessWidget {
+  const AuthShell({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const AutoRouter();
+  }
+}
