+++ lib/features/settings/presentation/shells/setting_shell.dart
@@ -1,0 +1,12 @@
+import 'package:auto_route/auto_route.dart';
+import 'package:flutter/material.dart';
+
+@RoutePage()
+class SettingShell extends StatelessWidget {
+  const SettingShell({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const AutoRouter();
+  }
+}
